@@ -8,7 +8,7 @@ class NewNormalizerEntryPoint(NormalizerEntryPoint):
     def load(self):
         from {{cookiecutter.module_name}}.normalizers.normalizer import NewNormalizer
 
-        return NewNormalizer(**self.dict())
+        return NewNormalizer(**self.model_dump())
 
 
 normalizer_entry_point = NewNormalizerEntryPoint(
